Change zellij config to non-colliding mode
There is a configuration that is meant to be non-colliding
where you default to locked and then unlock to use most commands
but some are still directly available.

To get this mode you have to overrite your config file because it isn't
just a setting. It is a whole bunch of configs